+++ pkgs/code_builder/lib/src/specs/expression/literal.dart
@@ -41,12 +41,17 @@ Expression literalNum(num value) => LiteralExpression._('$value');
 
 /// Create a literal expression from a string [value].
 ///
-/// Returns an expression that will evaluate to a String  containing exactly the
-/// same content as [value]. The literal may use single or double quotes, and
-/// may be raw, depending on the content. All disallowed characters are
-/// automatically escaped.
-Expression literalString(String value) =>
-    LiteralExpression._(_escapeString(value));
+/// Returns an expression for a string formatted `'<value>'`.
+///
+/// If [raw] is `true` returns an expression that will evaluate to a String
+/// containing exactly the same content as [value]. The literal may use single
+/// or double quotes, and may not actually be markedraw, depending on the
+/// content. All disallowed characters are automatically escaped.
+Expression literalString(String value, {bool raw = false}) {
+  if (raw) return LiteralExpression._(_escapeString(value));
+  final escaped = value.replaceAll('\'', '\\\'').replaceAll('\n', '\\n');
+  return LiteralExpression._("${raw ? 'r' : ''}'$escaped'");
+}
 
 String _escapeString(String value) {
   final original = value;